Remove check for lock when converting strings into IDs
If manufacturer of a computer is locked, this check prevents proper
build of SQL query, and also generates a PHP warning.
It also prevents display in GLPi of the value reported by the
FusionInventory Agent next to the locked value (tab locks (fields)).

The lock still operates and the field is not altered if its content
differs from the detected value bu the Agent.